+++ runtime/opts/opts_linux.go
@@ -14,7 +14,7 @@
    limitations under the License.
 */
 
-package containerd
+package opts
 
 import (
 	"context"
